+++ src/linux/PeripheralBase.cpp
@@ -22,6 +22,10 @@ PeripheralBase::~PeripheralBase() {
     _cleanup_characteristics();
 }
 
+void* PeripheralBase::underlying() const {
+    return device_.get();
+}
+
 std::string PeripheralBase::identifier() { return device_->name(); }
 
 BluetoothAddress PeripheralBase::address() { return device_->address(); }